// src/app/admin/appointments/type.ts
// types.ts


export interface Appointment {
    eventType: any;
    address: string;
    church?: string;
    createdAt: any;
    date: string;
    email: string;
    event: string;
    id: string;
    message: string;
    name: string;
    numberOfGuests: number;
    phone: string;
    receptionArea?: string;
    status: string;
    time: string;
}